+++ Planarian.Web/src/Modules/Import/Models/FailedCsvRecord.ts
@@ -1,0 +1,5 @@
+export interface FailedCsvRecord<TRecord> {
+  caveCsvModel: TRecord;
+  rowNumber: number;
+  reason: string;
+}
